tutorials/serverless-cloudrun-apigw-firebase/app/backend/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
gevent==21.1.2
google-api-core==1.26.3
google-auth==1.28.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.36.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
rsa==4.7.2
six==1.15.0
tldextract==3.1.0
urllib3==1.26.4
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0
